# OpenAPI.yaml
openapi: 3.0.0
info:
  title: SyncYomi Server API
  description: SyncYomi Server API
  contact:
    name: Kaiser
    email: kaiserbh@proton.me
    url: https://github.com/flurbudurbur/Shiori
  version: 1.0.0
servers:
  - url: http://localhost:8282/api
paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: User login
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '204':
          description: Login successful
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: User logout
      operationId: logoutUser
      responses:
        '204':
          description: Logout successful
        '500':
          description: Internal server error
  /auth/onboard:
    get:
      tags:
        - Authentication
      summary: Check if onboarding is possible
      description: Check if onboarding is possible
      operationId: canOnboard
      responses:
        '204':
          description: Onboarding is possible
        '403':
          description: Onboarding is unavailable
        '500':
          description: Internal server error
    post:
      tags:
        - Authentication
      summary: User onboarding
      description: User onboarding
      operationId: onboardUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '204':
          description: Onboarding successful
        '400':
          description: Bad request
        '403':
          description: Forbidden
        '500':
          description: Internal server error
  /auth/validate:
    get:
      tags:
        - Authentication
      summary: Validate if a user is authenticated
      description: Validate if a user is authenticated
      operationId: validateUser
      responses:
        '204':
          description: User is authenticated
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
...
  /logs/files:
    get:
      summary: Get log files
      description: Get log files
      operationId: getLogFiles
      tags:
        - Logs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogfilesResponse'
        default:
          description: Unexpected error
  /logs/files/{logFile}:
    get:
      summary: Download log file
      description: Download log file
      operationId: downloadLogFile
      tags:
        - Logs
      parameters:
        - name: logFile
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        default:
          description: Unexpected error
...
components:
  schemas:
...
    User:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
...
    LogfilesResponse:
      type: object
      properties:
        files:
          type: array
          items:
            $ref: '#/components/schemas/LogFile'
        count:
          type: integer
    LogFile:
      type: object
      properties:
        filename:
          type: string
        size_bytes:
          type: integer
        size:
          type: string
        updated_at:
          type: string
          format: date-time
...